feat(onboarding): rename widgets, move progress indicator to shared
